requirements.txt
streamlit
langchain
langchain_ollama
newspaper3k
beautifulsoup4
PyPDF2 